+++ src/java/org/apache/poi/ss/formula/atp/YearFracCalculator.java
@@ -27,7 +27,7 @@ import org.apache.poi.util.LocaleUtil;
 
 /**
  * Internal calculation methods for Excel 'Analysis ToolPak' function YEARFRAC()<br>
- *  
+ *
  * Algorithm inspired by www.dwheeler.com/yearfrac
  */
 final class YearFracCalculator {
@@ -61,7 +61,7 @@ final class YearFracCalculator {
 		int startDateVal = (int) Math.floor(pStartDateVal);
 		int endDateVal = (int) Math.floor(pEndDateVal);
 		if (startDateVal == endDateVal) {
-			// when dates are equal, result is zero 
+			// when dates are equal, result is zero
 			return 0;
 		}
 		// swap start and end if out of order
@@ -92,7 +92,7 @@ final class YearFracCalculator {
 		int date1day = startDate.day;
 		int date2day = endDate.day;
 
-		// basis zero has funny adjustments to the day-of-month fields when at end-of-month 
+		// basis zero has funny adjustments to the day-of-month fields when at end-of-month
 		if (date1day == LONG_MONTH_LEN && date2day == LONG_MONTH_LEN) {
 			date1day = SHORT_MONTH_LEN;
 			date2day = SHORT_MONTH_LEN;
@@ -155,7 +155,7 @@ final class YearFracCalculator {
 		int date2day = endDate.day;
 
 
-		// basis four has funny adjustments to the day-of-month fields when at end-of-month 
+		// basis four has funny adjustments to the day-of-month fields when at end-of-month
 		if (date1day == LONG_MONTH_LEN) {
 			date1day = SHORT_MONTH_LEN;
 		}
@@ -169,10 +169,10 @@ final class YearFracCalculator {
 
 	private static double calculateAdjusted(SimpleDate startDate, SimpleDate endDate, int date1day,
 			int date2day) {
-		double dayCount 
-			= (endDate.year - startDate.year) * 360
-			+ (endDate.month - startDate.month) * SHORT_MONTH_LEN
-			+ (date2day - date1day) * 1;
+		double dayCount
+			= (endDate.year - startDate.year) * 360.0
+			+ (endDate.month - startDate.month) * (double)SHORT_MONTH_LEN
+			+ (date2day - date1day) * 1.0;
 		return dayCount / 360;
 	}
 
@@ -223,7 +223,7 @@ final class YearFracCalculator {
 			}
 			return false;
 		}
-		
+
 		if (isLeapYear(end.year)) {
 			switch (end.month) {
 				case SimpleDate.JANUARY:
@@ -245,14 +245,14 @@ final class YearFracCalculator {
 	private static int dateDiff(long startDateMS, long endDateMS) {
 		long msDiff = endDateMS - startDateMS;
 
-		// some extra checks to make sure we don't hide some other bug with the rounding 
+		// some extra checks to make sure we don't hide some other bug with the rounding
 		int remainderHours = (int) ((msDiff % MS_PER_DAY) / MS_PER_HOUR);
 		switch (remainderHours) {
 			case 0:  // normal case
 				break;
 			case 1:  // transition from normal time to daylight savings adjusted
 			case 23: // transition from daylight savings adjusted to normal time
-				// Unexpected since we are using UTC_TIME_ZONE 
+				// Unexpected since we are using UTC_TIME_ZONE
 			default:
 				throw new RuntimeException("Unexpected date diff between " + startDateMS + " and " + endDateMS);
